move tests into it's own folder.
