+++ JAVMovieScraper/pom.xml
@@ -4,7 +4,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>groupId</groupId>
+    <groupId>moviescraper.doctord</groupId>
     <artifactId>JAVMovieScraper</artifactId>
     <version>1.0-SNAPSHOT</version>
 
@@ -13,12 +13,20 @@
             <groupId>org.hamcrest</groupId>
             <artifactId>hamcrest-all</artifactId>
             <version>1.3</version>
+            <scope>test</scope>
         </dependency>
 
         <dependency>
             <groupId>commons-io</groupId>
             <artifactId>commons-io</artifactId>
             <version>2.4</version>
+        </dependency>
+
+        <dependency>
+            <groupId>junit</groupId>
+            <artifactId>junit</artifactId>
+            <version>4.8.2</version>
+            <scope>test</scope>
         </dependency>
 
         <dependency>
@@ -90,9 +98,11 @@
     </dependencies>
 
     <build>
+        <testSourceDirectory>src-tests</testSourceDirectory>
+
         <resources>
             <resource>
-                <directory>./src/</directory>
+                <directory>${project.basedir}/src/</directory>
             </resource>
         </resources>
